+++ www/Library/content_widgets/Comments/Comments.php
@@ -6,30 +6,10 @@
  */
 namespace Library\content_widgets\Comments;
 
-use Library\basic\widgets\ViewObjectsList\ViewObjectsList,
-    Boolive\values\Rule;
+use Library\views\AutoWidgetList\AutoWidgetList;
 
-class Comments extends ViewObjectsList
+class Comments extends AutoWidgetList
 {
-    public function getInputRule()
-    {
-        return Rule::arrays(array(
-            'GET' => Rule::arrays(array(
-                'object' => Rule::entity()->required(), // объект, который отображать
-                ), Rule::any() // не удалять другие элементы
-            )), Rule::any() // не удалять другие элементы
-        );
-    }
-
-    public function canWork()
-    {
-        if ($result = parent::canWork()){
-            // По URL определяем объект и номер страницы
-            $this->_input['GET']['objects_list'] = $this->_input['GET']['object']->findAll();
-        }
-        return $result;
-    }
-
     public function work($v = array())
     {
         echo 'Comments';