+++ www/Library/content_widgets/Title/Title.php
@@ -6,21 +6,10 @@
  */
 namespace Library\content_widgets\Title;
 
-use Library\basic\widgets\Widget\Widget,
-    Boolive\values\Rule;
+use Library\views\Widget\Widget;
 
 class Title extends Widget
 {
-    public function getInputRule()
-    {
-        return Rule::arrays(array(
-            'GET' => Rule::arrays(array(
-                'object' => Rule::entity()->required(), // объект, который отображать
-                ), Rule::any() // не удалять другие элементы
-            )), Rule::any() // не удалять другие элементы
-        );
-    }
-
     public function work($v = array())
     {
         $v['value'] = $this->_input['GET']['object']->getValue();
@@ -32,7 +21,6 @@ class Title extends Widget
         }else{
             $v['parent_uri'] = '';
         }
-
         return parent::work($v);
     }
 }